Parameter creation
Added createFindSigFreqsParameterStruct, which will allow the user to
interactively define parameters. Also made small changes to
DiffCondsSignif that allow it to work with the parameter struct:
Changed comment to match the actual field value; made it such that
having params.neighborMat empty be the same thing as having it
undefined.