Added multi_channel and top-percentile features
Multi_Channel: will run across multiple channels and plot the
significant frequencies for each.
Top-percentile: allows user to specify a percent of data to use for
checking for significance (iteratively chooses the likeliest frequency
to be significant and then takes the X% of trials with the highest
spectral power at that frequency).